Clean up and format
